Use build name as default context and description.
If context is left blank now defaults to the environmental
variable "BUILD_JOB_NAME" which is passed as metadata by
Concourse and contains the job name.

If description is left blank now defaults to a string
specifying the build name and the build number.

This change, together with the use of anchors in the
pipeline configuration file, can significantly reduce
the amount of duplication required.

Signed-off-by: Alessandro Degano <alessandro.degano@pix4d.com>